EndFragment implemented

+++ app/src/main/java/com/palmdev/learn_math/data/model/ResultExercise.kt
@@ -5,6 +5,6 @@ import com.palmdev.learn_math.utils.Operation
 data class ResultExercise(
     val id: Long? = null,
     val operation: Operation,
-    val correctAnswersPercent: Float, // 0.5f -> 50%
-    val avgAnswerTime: Float // 3.5f -> 3.5s
+    val correctAnswersPercent: Int, // 0.5f -> 50%
+    val avgAnswerTime: Double // 3.50 -> 3.5s
 )
