+++ app/src/main/java/com/palmdev/learn_math/data/model/ResultExerciseEntity.kt
@@ -9,6 +9,6 @@ data class ResultExerciseEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long? = null,
     val operation: String,
-    val correctAnswersPercent: Float,
-    val avgAnswerTime: Float // seconds
+    val correctAnswersPercent: Int,
+    val avgAnswerTime: Double // seconds
 )
